refactor: Search에서 Dto디렉토리 하부의 request response 디렉토리 제거
