+++ src/main/java/com/kustacks/kuring/notice/business/NoticeService.java
@@ -9,7 +9,7 @@ import com.kustacks.kuring.notice.common.dto.NoticeDto;
 import com.kustacks.kuring.notice.common.dto.NoticeListResponse;
 import com.kustacks.kuring.notice.domain.Notice;
 import com.kustacks.kuring.notice.domain.NoticeRepository;
-import com.kustacks.kuring.search.common.dto.response.NoticeSearchDto;
+import com.kustacks.kuring.search.common.dto.NoticeSearchDto;
 import org.springframework.beans.factory.annotation.Value;
 import org.springframework.data.domain.PageRequest;
 import org.springframework.stereotype.Service;
