+++ src/main/java/com/kustacks/kuring/notice/common/dto/NoticeLookupResponse.java
@@ -1,6 +1,6 @@
 package com.kustacks.kuring.notice.common.dto;
 
-import com.kustacks.kuring.search.common.dto.response.NoticeSearchDto;
+import com.kustacks.kuring.search.common.dto.NoticeSearchDto;
 import lombok.AllArgsConstructor;
 import lombok.Getter;
 import lombok.NoArgsConstructor;
