+++ src/main/java/com/kustacks/kuring/search/common/dto/NoticeWebSocketResponseDto.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.search.common.dto.response;
+package com.kustacks.kuring.search.common.dto;
 
 import lombok.Getter;
