+++ src/main/java/com/kustacks/kuring/search/common/dto/WebSocketResponseDto.java
@@ -1,4 +1,4 @@
-package com.kustacks.kuring.search.common.dto.response;
+package com.kustacks.kuring.search.common.dto;
 
 import com.kustacks.kuring.common.dto.ResponseDto;
 import lombok.AccessLevel;
